Stereo- and mono-camera mode, multiple windows
You can choose between a stereo and a mono camera mode for the release renderer.
In the release window, press M for mono and S for stereo to switch modes.
Press the left and right arrow keys to change the eyeSperation value.
Press the down and up arrow keys to change the zZeroParallax value.

Two windows, one for the release rendering and one for a debug rendering, will be created.