Deferred Shading, Stencil Shadow Volumes
Deferred Shading is used to minimize the amount of pixels for which the lighting must be performed.
Stencil Shadow Volumes are used for geometry based (self-)shadows. For this technique the geometry is required to be a triangle list with adjacencies.

Optimization to be done: Limit the amount of geometry for the shadow volumes to those in the point light radius.