Added SSAO, Gaussian Blur, Directional Light Technique (all post processing)
Screen Space Ambient Occlusion is calculated for the scene, blured with a gaussian filter and applied to final image.
There is one directional light, currently beaming a grey color in the direction (0, -2, -1) in view space.
The old shadow volume and light technique became obsolete with this word but they are not yet deleted. Should be discussed with the team first.

Cleaned up code a bit.
Separated geometry buffer and geometry technique.
Removed unnecessary includes. Should speed up compiling a little.
Empty constructors need not be called explicitly.
Made most one-liner functions inline and moved them to the header files.